Add =~ and !~ operators, refactor tests, fix bugs
Regex operators
---------------

The =~ operator can be used like this:

    x =~ "^(foo|bar|baz)*bingo$"

the following values for x will match, for example:

    foofoobingo
    barfoobarbazbingo
    ...

invert the sense of the match with the !~ operator.

Refactor tests
--------------

The test-data/ directory contains compiled zerkel test cases, compiled
with older versions of the zerkel compiler. The test harness now runs
tests against all previous versions that should be compatible with this
version, too, so we can detect failure of backward compatibility.

Bug fixes
---------

- Variables names can now contain the $ character in any place. Previous
  behavior allowed $ only at the start of the variable name.

- Variable property names (eg. the "bar" in foo.bar) can now contain the
  $ character. Previously they could not.

- Variable property reference syntax is now robust. For example the
  following no longer no longer parse:

      foo..bar
      foo.bar.

- The "reversed" syntax for the LIKE operator is no longer supported
  (eg. "*foo*" LIKE "hello foo world"). This was always broken due to a
  bug in the matcher function, causing the reversed syntax to always
  result in the query evaluating to true. So instead of fixing this bug
  we simply don't attempt to support this (previously undocumented)
  behavior.